+++ jgclark.Dashboard/src/dataGenerationDays.js
@@ -14,14 +14,13 @@ import {
   getNotePlanSettings,
   getOpenItemParasForTimePeriod,
   getStartTimeFromPara,
-  makeDashboardParas,
 } from './dashboardHelpers'
 import { openTodayItems, refTodayItems, openTomorrowParas, refTomorrowParas, openYesterdayParas, refYesterdayParas } from './demoData'
-import { getTodaysDateHyphenated, getTodaysDateUnhyphenated } from '@helpers/dateTime'
+import { getTodaysDateUnhyphenated } from '@helpers/dateTime'
 import { clo, JSP, logDebug, logError, logInfo, logTimer, logWarn, timer } from '@helpers/dev'
 import { toNPLocaleDateString } from '@helpers/NPdateTime'
 import { getHeadingsFromNote } from '@helpers/NPnote'
-import { getCurrentTimeBlockPara, getTimeBlockDetails, isActiveOrFutureTimeBlockPara } from '@helpers/timeblocks'
+import { isActiveOrFutureTimeBlockPara } from '@helpers/timeblocks'
 import { isOpen } from '@helpers/utils'
 
 //--------------------------------------------------------------------
@@ -200,6 +199,7 @@ export function getTodaySectionData(config: TDashboardSettings, useDemoData: boo
           postActionRefresh: ['DT', 'DO'], // refresh 2 sections afterwards
         },
       ],
+      isReferenced: false,
     }
     // clo(section)
     sections.push(section)
@@ -242,6 +242,7 @@ export function getTodaySectionData(config: TDashboardSettings, useDemoData: boo
         sectionItems: items,
         generatedDate: new Date(), // Note: this often gets stringified to a string, but isn't underneath
         actionButtons: [],
+        isReferenced: true,
       }
       sections.push(section)
     }
@@ -285,6 +286,7 @@ export function getTodaySectionData(config: TDashboardSettings, useDemoData: boo
         sectionItems: timeBlockItems,
         generatedDate: new Date(),
         actionButtons: [],
+        isReferenced: false,
       }
       logTimer('getTodaySectionData', timer, `- found ${String(timeBlockItems.length)} timeblock items from ${filenameDateStr}`)
       sections.push(section)
@@ -390,6 +392,7 @@ export function getYesterdaySectionData(config: TDashboardSettings, useDemoData:
           postActionRefresh: ['DT', 'DY'], // refresh 2 sections afterwards
         },
       ],
+      isReferenced: false,
     }
     sections.push(section)
 
@@ -430,6 +433,7 @@ export function getYesterdaySectionData(config: TDashboardSettings, useDemoData:
         sectionItems: items,
         generatedDate: new Date(),
         actionButtons: [],
+        isReferenced: true,
       }
       sections.push(section)
     }
@@ -521,6 +525,7 @@ export function getTomorrowSectionData(config: TDashboardSettings, useDemoData: 
       sectionItems: items,
       generatedDate: new Date(),
       actionButtons: [],
+      isReferenced: false,
     }
     // clo(section, 'Tomorrow section')
     sections.push(section)
@@ -562,6 +567,7 @@ export function getTomorrowSectionData(config: TDashboardSettings, useDemoData: 
         sectionItems: items,
         generatedDate: new Date(),
         actionButtons: [],
+        isReferenced: true,
       }
       // clo(section, `>Tomorrow section`)
       sections.push(section)
@@ -572,117 +578,5 @@ export function getTomorrowSectionData(config: TDashboardSettings, useDemoData: 
   } catch (error) {
     logError('getTomorrowSectionData', `ERROR: ${error.message}`)
     return []
-  }
-}
-
-/**
- * Get the current time block paras from Today's note if it exists.
- * Ignore any time block paras that are done or cancelled.
- * TODO: Make this cover time blocks for today scheduled from regular notes. Note: these are calculated normally about the same time as this section. So perhaps roll the two together and generate two sections?
- * @param {TDashboardSettings} config
- * @param {boolean} useDemoData?
- * @returns {Array<TSection>}
- */
-export function getTimeBlockSectionData(_config: TDashboardSettings, useDemoData: boolean = false): TSection {
-  try {
-    const sectionNumStr = '116' // FIXME: revert later
-    const thisSectionCode = '_TB' // FIXME: revert later
-    const items: Array<TSectionItem> = []
-    const NPSettings = getNotePlanSettings()
-    const mustContainString = NPSettings.timeblockMustContainString
-    const thisFilename = `${getTodaysDateUnhyphenated()}.${NPSettings.defaultFileExtension}`
-    const filenameDateStr = moment().format('YYYYMMDD') // use Moment so we can work on local time and ignore TZs
-    const currentDailyNote = DataStore.calendarNoteByDateString(filenameDateStr)
-    logInfo('getTimeBlockSectionData', `--------- Gathering${useDemoData ? ' DEMO' : ''} time blocks from ${filenameDateStr} with mCS ${mustContainString} ----------`)
-    const startTime = new Date() // for timing only
-    let timeblockPara: ?TParagraph
-
-    if (useDemoData) {
-      const fakeTodayNoteParagraphs: Array<TParagraphForDashboard> = []
-      for (const item of openTodayItems) {
-        if (item.para) fakeTodayNoteParagraphs.push(item.para)
-      }
-      // $FlowFixMe[prop-missing]
-      // $FlowFixMe[incompatible-type]
-      const fakeTodayNote: TNote = {
-        // $FlowFixMe[incompatible-type]
-        // $FlowFixMe[prop-missing]
-        // $FlowFixMe[incompatible-exact]
-        paragraphs: fakeTodayNoteParagraphs,
-        type: 'Calendar',
-        title: getTodaysDateHyphenated(),
-        filename: `${filenameDateStr}.md`,
-      }
-      clo(fakeTodayNote, `fakeTodayNote`)
-      timeblockPara = getCurrentTimeBlockPara(fakeTodayNote, true, mustContainString)
-    } else {
-      // Get list of open tasks/checklists from current daily note (if it exists)
-      if (currentDailyNote) {
-        timeblockPara = getCurrentTimeBlockPara(currentDailyNote, true, mustContainString)
-      } else {
-        logDebug('getTimeBlockSectionData', `No daily note found using filename '${filenameDateStr}'`)
-      }
-    }
-
-    if (!timeblockPara) {
-      logDebug('getTimeBlockSectionData', `-> none`)
-    } else {
-      const timeBlockParaContent = timeblockPara?.content ?? ''
-      const [timeBlockString, contentWithoutTimeBlock] = getTimeBlockDetails(timeBlockParaContent, mustContainString)
-      logDebug('getTimeBlockSectionData', `-> ${timeBlockString} / ${contentWithoutTimeBlock}`)
-
-      // write item for section
-      const thisID = `${sectionNumStr}-0`
-      // I was trying to reorder the display of the timeblock para content, but that makes life very difficult for editing or checking off time block items
-      const thisDParas = makeDashboardParas([timeblockPara])
-      // // change text so that time details are always at the front
-      // // If this timeblock is on a task or checklist line, leave the type alone, otherwise make it type 'timeblock'. This affects StatusIcon.
-      // thisDPara.content = `${timeBlockString} ${contentWithoutTimeBlock}`
-      // // handle priority markers
-      // if (thisDPara.priority > 0) {
-      //   const priorityMarker = ['!', '!!', '!!!', '>>'][thisDPara.priority - 1]
-      //   thisDPara.content = `${priorityMarker} ${thisDPara.content}`
-      // }
-      if (thisDParas.length > 0) {
-        const thisDPara = thisDParas[0]
-        if (thisDPara) {
-          const itemTypeToUse = thisDPara.type === 'open' || thisDPara.type === 'checklist' ? thisDPara.type : 'timeblock'
-          const thisSectionItemObject = { ID: thisID, itemType: itemTypeToUse, para: thisDPara }
-          // $FlowFixMe[incompatible-call]
-          items.push(thisSectionItemObject)
-        } else {
-          logDebug('getTimeBlockSectionData', `Can't fully show time block as this is DEMO data.`)
-        }
-      } else {
-        if (useDemoData) {
-          logDebug('getTimeBlockSectionData', `Can't fully show time block as this is DEMO data.`)
-        } else {
-          logWarn('getTimeBlockSectionData', `Couldn't find thisDPara to match '${timeBlockParaContent}'. Doesn't `)
-        }
-      }
-    }
-
-    const section: TSection = {
-      ID: sectionNumStr,
-      name: 'Current time block',
-      showSettingName: 'showTimeBlockSection',
-      sectionCode: thisSectionCode,
-      description: '', //`current time block`,
-      FAIconClass: 'fa-light fa-calendar-clock',
-      sectionTitleColorPart: 'sidebarYearly',
-      sectionFilename: thisFilename,
-      // sectionItems: items,
-      sectionItems: [],
-      generatedDate: new Date(),
-      actionButtons: [],
-    }
-    // clo(section)
-    logTimer('getTimeBlockSectionData', startTime, `- found Current Time Block from ${filenameDateStr}`)
-
-    return section
-  } catch (error) {
-    logError(`getTimeBlockSectionData`, error.message)
-    // $FlowFixMe[incompatible-return]
-    return null
   }
 }